added confirm message to delete bug and fixed navigation bug

+++ src/app/search-bug/search-bug.component.ts
@@ -81,14 +81,17 @@ export class SearchBugComponent implements OnInit {
       }, error => alert("Error occurred"));
     }
   }
-  deleteBug(id:string,index:number){
-    const observable = this.bugService.deleteBug(id);
-    observable.subscribe(response => {
-      alert("Bug deleted");
-      this.bugArray.splice(index,1);
-    }, error => {
-      alert("Error occurred");
-    });
+  deleteBug(id: string, index: number, name:string) {
+    if (confirm("Delete bug : " + name)) {
+      const observable = this.bugService.deleteBug(id);
+      observable.subscribe(response => {
+        alert("Bug deleted");
+        this.bugArray.splice(index, 1);
+      }, error => {
+        alert("Error occurred");
+      });
+    }
+
   }
 
   ngOnInit(): void {
